SSR bootstrap most recent site into InterimOmnibar (DOTMSD-1167)
Extract shared HMAC auth helper from user-bootstrap into
bootstrap-auth.ts. Add site-bootstrap.ts that fetches the user's
most recent site server-side (URL slug > preferences > primary_blog)
in parallel with user bootstrap. Inject the site into the SSR-rendered
InterimOmnibar and window.mostRecentSite for hydration.

Also extract SITE_FIELDS/SITE_OPTIONS from api-core fetchers into
a standalone constants.ts so they can be imported server-side without
pulling in browser-only wpcom dependencies.
