+++ config/dashboard-development.json
@@ -28,7 +28,7 @@
 		"ciab/allow-domain-features": true,
 		"cookie-banner": false,
 		"rum-tracking/logstash": true,
-		"dashboard/omnibar": false,
+		"dashboard/omnibar": true,
 		"dashboard/wp-beta-program": true,
 		"dashboard/v2": false,
 		"dev/auth-helper": true,
@@ -47,7 +47,8 @@
 		"site-spec": true,
 		"wordpress-ai-tools": true,
 		"two-factor/enhanced-security": true,
-		"wpcom-user-bootstrap": false
+		"wpcom-user-bootstrap": true,
+		"wpcom-bootstrap-most-recent-site": true
 	},
 	"enable_all_sections": false,
 	"sections": {
