+++ config/dashboard-production.json
@@ -44,7 +44,8 @@
 		"site-spec": true,
 		"wordpress-ai-tools": true,
 		"two-factor/enhanced-security": true,
-		"wpcom-user-bootstrap": true
+		"wpcom-user-bootstrap": true,
+		"wpcom-bootstrap-most-recent-site": true
 	},
 	"enable_all_sections": false,
 	"sections": {
